added a script for transcribing wav files to midi (#7)
* Tab separated headers for tsv

* added a script for transcribing wav files to midi

* Bump pillow from 5.3.0 to 6.2.0

Bumps [pillow](https://github.com/python-pillow/Pillow) from 5.3.0 to 6.2.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/5.3.0...6.2.0)


* using git+https url for mir_eval

* fixed the broken link to the TF implementation

* some cosmetic changes

* train_on option for training on MAPS instead

* Rename transcribe_wav.py to transcribe.py

* Update decoding.py

* changed CLI option to take multiple files

+++ onsets_and_frames/utils.py
@@ -71,8 +71,8 @@ def save_pianoroll(path, onsets, frames, onset_threshold=0.5, frame_threshold=0.
     frame_threshold: float
     zoom: int
     """
-    onsets = (1 - (onsets.t() > onset_threshold)).cpu()
-    frames = (1 - (frames.t() > frame_threshold)).cpu()
+    onsets = (1 - (onsets.t() > onset_threshold).to(torch.uint8)).cpu()
+    frames = (1 - (frames.t() > frame_threshold).to(torch.uint8)).cpu()
     both = (1 - (1 - onsets) * (1 - frames))
     image = torch.stack([onsets, frames, both], dim=2).flip(0).mul(255).numpy()
     image = Image.fromarray(image, 'RGB')
